Forest plots for brmsfit objects

This function draws a forest plot for a brmsfit model; especially useful for random-effects meta-analytic models.

forest(model, pars = NA, level = 0.95, av_name = "Average", sort = TRUE,
  show_data = FALSE, col_ridge = NA, fill_ridge = "grey75",
  density = TRUE, text = TRUE, rel_min_height = 0.01, scale = 0.9,
  digits = 2, theme_forest = TRUE)

Arguments

model
A brmsfit object.

pars
Parameters to plot, defaults to all (NA).

level
The "Confidence" level for the Credible Intervals. Defaults to 0.95.

av_name
Name of average parameter (e.g. `"Meta-Analytic\nestimate"`)

sort
Logical; whether to sort the estimates in ascending order of magnitude from bottom to top. Defaults to FALSE.

show_data
Logical; whether to show the observed effect size and standard error below the meta-analytic estimates. Defaults to FALSE.

col_ridge
String; color for the outlines of the densities. Default: NA.

fill_ridge
String; color to fill the densities. Defaults to "grey65".

density
Logical; should densities be shown?

text
Logical; should estimates' numerical summaries be included?

rel_min_height
Passed to [ggridges::geom_density_ridges()].

scale
Passed to [ggridges::geom_density_ridges()].

digits
Digits to display in numerical summaries.

theme_forest
Use [brmstools::theme_forest()] ggplot2 theme?

Value

a ggplot